Update app-service-api-arm-new-gateway-provision.md
Hi Team,
Added app settings to configure authentication for AAD identity for the app gateway, it took me a while to figure that out, had to reverse engineer that from Portal by intercepting the traffic. Hope this will save someone's day. 
It may be that adding an authentication will make it harder for people to test APIs (first timers specially) so feel free to add it in the description if that's more suitable.

Thanks,
Suneet Nangia
CSA, MSUK